rename "shuffleView" to "get_disk_positions"
